+++ delphi-frontend/src/main/java/au/com/integradev/delphi/antlr/ast/node/AttributeListNodeImpl.java
@@ -20,10 +20,15 @@ package au.com.integradev.delphi.antlr.ast.node;
 
 import au.com.integradev.delphi.antlr.ast.visitors.DelphiParserVisitor;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.antlr.runtime.Token;
 import org.sonar.plugins.communitydelphi.api.ast.AttributeGroupNode;
 import org.sonar.plugins.communitydelphi.api.ast.AttributeListNode;
 import org.sonar.plugins.communitydelphi.api.ast.AttributeNode;
+import org.sonar.plugins.communitydelphi.api.symbol.declaration.NameDeclaration;
+import org.sonar.plugins.communitydelphi.api.symbol.declaration.TypeNameDeclaration;
+import org.sonar.plugins.communitydelphi.api.type.Type;
+import org.sonar.plugins.communitydelphi.api.type.TypeFactory;
 
 public final class AttributeListNodeImpl extends DelphiNodeImpl implements AttributeListNode {
   public AttributeListNodeImpl(Token token) {
@@ -47,5 +52,25 @@ public final class AttributeListNodeImpl extends DelphiNodeImpl implements Attri
   @Override
   public List<AttributeNode> getAttributes() {
     return findDescendantsOfType(AttributeNode.class);
+  }
+
+  @Override
+  public List<Type> getAttributeTypes() {
+    return getAttributes().stream()
+        .map(AttributeNode::getTypeNameOccurrence)
+        .map(
+            occurrence -> {
+              if (occurrence == null) {
+                return TypeFactory.unknownType();
+              }
+
+              NameDeclaration declaration = occurrence.getNameDeclaration();
+              if (!(declaration instanceof TypeNameDeclaration)) {
+                return TypeFactory.unknownType();
+              }
+
+              return ((TypeNameDeclaration) declaration).getType();
+            })
+        .collect(Collectors.toList());
   }
 }